Update footer heart symbol
Change the heart symbol into an emoji for
accessibility. I scanned the site with an
accessibility tool and it notified me that the
color on the "<3" text representation of a heart
was of poor contrast. I made the contrast higher
but disliked the color, so I opted for an emoji
instead. An emoji will be read aloud to a screen
reader, has an easy to recognize shape, and
has the color that I would like. It's a mix of
dark purple and red, so it does contain some
high contrast elements to it as well.